ENH: Minor readability improvement in CTest output
When ctest --build-and-test runs the --test-command its output did not
quote the arguments of the command being tested making it difficult to
read.  This adds the quotes.  This also changes the wording of the
failure case to not sound like CTest could not run the executable when
in fact it ran and returned failure.