ENH: Clean per-config export files during install
When installing the main export file the install tree may be dirty.  If
out-dated per-config files exist they may break the newly installed main
file which when it globs them.  This teaches the installation script to
detect when it is about to replace the main export file with a different
one and cleans out any existing per-config files.